undo the sage_jupyter.py file split from sage_salvus.py
